Update issue 256
アクティビティポートレットを追加しました。
スケジュールポートレットのアクティビティを登録しました。
ブログポートレットのアクティビティを登録しました。

git-svn-id: http://aipo.googlecode.com/svn/trunk@741 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ oms/src/main/java/com/aimluck/eip/cayenne/om/social/auto/_ActivityMap.java
@@ -32,11 +32,11 @@ public class _ActivityMap extends org.apache.cayenne.CayenneDataObject {
 
     public static final String ID_PK_COLUMN = "ID";
 
-    public void setIsRead(Short isRead) {
+    public void setIsRead(Integer isRead) {
         writeProperty("isRead", isRead);
     }
-    public Short getIsRead() {
-        return (Short)readProperty("isRead");
+    public Integer getIsRead() {
+        return (Integer)readProperty("isRead");
     }
